+++ src/main/java/root/git_turl/domain/member/entity/Member.java
@@ -5,9 +5,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import root.git_turl.domain.enums.JobType;
-import root.git_turl.domain.enums.Status;
-import root.git_turl.domain.enums.TechStack;
+import root.git_turl.domain.member.enums.JobType;
+import root.git_turl.domain.member.enums.Status;
+import root.git_turl.domain.member.enums.TechStack;
 import root.git_turl.global.entity.BaseEntity;
 
 import java.util.ArrayList;
