+++ src/main/java/root/git_turl/domain/member/enums/JobType.java
@@ -1,4 +1,4 @@
-package root.git_turl.domain.enums;
+package root.git_turl.domain.member.enums;
 
 public enum JobType {
     FRONTEND, BACKEND, AI
