+++ src/main/java/root/git_turl/domain/member/enums/Status.java
@@ -1,4 +1,4 @@
-package root.git_turl.domain.enums;
+package root.git_turl.domain.member.enums;
 
 public enum Status {
     ACTIVATE, INACTIVATE
